x86/tdx: Move FRAME_BEGIN/END to TDX_MODULE_CALL asm macro
Currently, the TDX_MODULE_CALL asm macro and the __tdx_module_call()
take registers directly as input and a 'struct tdx_module_output' as
optional output.  This is different from the __tdx_hypercall(), which
simply uses a structure to carry all input/output.  There's no point to
leave __tdx_module_call() complicated as it is.

As a preparation to simplify the __tdx_module_call() to make it look
like __tdx_hypercall(), move FRAME_BEGIN/END and RET from the
__tdx_module_call() to the TDX_MODULE_CALL assembly macro.  This also
allows more implementation flexibility of the assembly inside the
TDX_MODULE_CALL macro, e.g., allowing putting an _ASM_EXTABLE() after
the main body of the assembly.

This is basically based on Peter's code.

Cc: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>
Cc: Dave Hansen <dave.hansen@linux.intel.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Suggested-by: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Kai Huang <kai.huang@intel.com>